#981 Add highest derivative values in general log with the filter DEBUG

+++ dynawo/sources/Solvers/AlgebraicSolvers/DYNSolverKINAlgRestoration.cpp
@@ -230,9 +230,9 @@ SolverKINAlgRestoration::evalF_KIN(N_Vector yy, N_Vector rr, void *data) {
 
   const int nbErr = 10;
   Trace::debug() << DYNLog(KinLargestErrors, nbErr) << Trace::endline;
-  vector<std::pair<double, int> > fErr;
-  for (unsigned int i = 0; i < solv->indexF_.size(); ++i)
-    fErr.push_back(std::pair<double, int>(solv->F_[solv->indexF_[i]], i));
+  vector<std::pair<double, size_t> > fErr;
+  for (size_t i = 0; i < solv->indexF_.size(); ++i)
+    fErr.push_back(std::pair<double, size_t>(solv->F_[solv->indexF_[i]], i));
   SolverCommon::printLargestErrors(fErr, model, nbErr);
 #endif
   return (0);
